fix onlinetest bugs and delete sone unused table thead
