+++ src/hibernatePOJO/PowerxbMonitor.java
@@ -610,6 +610,12 @@ public class PowerxbMonitor {
     private Float i1Va50;
     private Float i2Va50;
     private Float i3Va50;
+    private Float thdu1;
+    private Float thdu2;
+    private Float thdu3;
+    private Float thdi1;
+    private Float thdi2;
+    private Float thdi3;
 
     @Id
     @Column(name = "xbid", nullable = false)
@@ -7255,5 +7261,53 @@ public class PowerxbMonitor {
     public int hashCode() {
 
         return Objects.hash(xbid, did, time, u1Xb1, u2Xb1, u3Xb1, u1Va1, u2Va1, u3Va1, i1Xb1, i2Xb1, i3Xb1, i1Va1, i2Va1, i3Va1, u1Xb2, u2Xb2, u3Xb2, u1Va2, u2Va2, u3Va2, i1Xb2, i2Xb2, i3Xb2, i1Va2, i2Va2, i3Va2, u1Xb3, u2Xb3, u3Xb3, u1Va3, u2Va3, u3Va3, i1Xb3, i2Xb3, i3Xb3, i1Va3, i2Va3, i3Va3, u1Xb4, u2Xb4, u3Xb4, u1Va4, u2Va4, u3Va4, i1Xb4, i2Xb4, i3Xb4, i1Va4, i2Va4, i3Va4, u1Xb5, u2Xb5, u3Xb5, u1Va5, u2Va5, u3Va5, i1Xb5, i2Xb5, i3Xb5, i1Va5, i2Va5, i3Va5, u1Xb6, u2Xb6, u3Xb6, u1Va6, u2Va6, u3Va6, i1Xb6, i2Xb6, i3Xb6, i1Va6, i2Va6, i3Va6, u1Xb7, u2Xb7, u3Xb7, u1Va7, u2Va7, u3Va7, i1Xb7, i2Xb7, i3Xb7, i1Va7, i2Va7, i3Va7, u1Xb8, u2Xb8, u3Xb8, u1Va8, u2Va8, u3Va8, i1Xb8, i2Xb8, i3Xb8, i1Va8, i2Va8, i3Va8, u1Xb9, u2Xb9, u3Xb9, u1Va9, u2Va9, u3Va9, i1Xb9, i2Xb9, i3Xb9, i1Va9, i2Va9, i3Va9, u1Xb10, u2Xb10, u3Xb10, u1Va10, u2Va10, u3Va10, i1Xb10, i2Xb10, i3Xb10, i1Va10, i2Va10, i3Va10, u1Xb11, u2Xb11, u3Xb11, u1Va11, u2Va11, u3Va11, i1Xb11, i2Xb11, i3Xb11, i1Va11, i2Va11, i3Va11, u1Xb12, u2Xb12, u3Xb12, u1Va12, u2Va12, u3Va12, i1Xb12, i2Xb12, i3Xb12, i1Va12, i2Va12, i3Va12, u1Xb13, u2Xb13, u3Xb13, u1Va13, u2Va13, u3Va13, i1Xb13, i2Xb13, i3Xb13, i1Va13, i2Va13, i3Va13, u1Xb14, u2Xb14, u3Xb14, u1Va14, u2Va14, u3Va14, i1Xb14, i2Xb14, i3Xb14, i1Va14, i2Va14, i3Va14, u1Xb15, u2Xb15, u3Xb15, u1Va15, u2Va15, u3Va15, i1Xb15, i2Xb15, i3Xb15, i1Va15, i2Va15, i3Va15, u1Xb16, u2Xb16, u3Xb16, u1Va16, u2Va16, u3Va16, i1Xb16, i2Xb16, i3Xb16, i1Va16, i2Va16, i3Va16, u1Xb17, u2Xb17, u3Xb17, u1Va17, u2Va17, u3Va17, i1Xb17, i2Xb17, i3Xb17, i1Va17, i2Va17, i3Va17, u1Xb18, u2Xb18, u3Xb18, u1Va18, u2Va18, u3Va18, i1Xb18, i2Xb18, i3Xb18, i1Va18, i2Va18, i3Va18, u1Xb19, u2Xb19, u3Xb19, u1Va19, u2Va19, u3Va19, i1Xb19, i2Xb19, i3Xb19, i1Va19, i2Va19, i3Va19, u1Xb20, u2Xb20, u3Xb20, u1Va20, u2Va20, u3Va20, i1Xb20, i2Xb20, i3Xb20, i1Va20, i2Va20, i3Va20, u1Xb21, u2Xb21, u3Xb21, u1Va21, u2Va21, u3Va21, i1Xb21, i2Xb21, i3Xb21, i1Va21, i2Va21, i3Va21, u1Xb22, u2Xb22, u3Xb22, u1Va22, u2Va22, u3Va22, i1Xb22, i2Xb22, i3Xb22, i1Va22, i2Va22, i3Va22, u1Xb23, u2Xb23, u3Xb23, u1Va23, u2Va23, u3Va23, i1Xb23, i2Xb23, i3Xb23, i1Va23, i2Va23, i3Va23, u1Xb24, u2Xb24, u3Xb24, u1Va24, u2Va24, u3Va24, i1Xb24, i2Xb24, i3Xb24, i1Va24, i2Va24, i3Va24, u1Xb25, u2Xb25, u3Xb25, u1Va25, u2Va25, u3Va25, i1Xb25, i2Xb25, i3Xb25, i1Va25, i2Va25, i3Va25, u1Xb26, u2Xb26, u3Xb26, u1Va26, u2Va26, u3Va26, i1Xb26, i2Xb26, i3Xb26, i1Va26, i2Va26, i3Va26, u1Xb27, u2Xb27, u3Xb27, u1Va27, u2Va27, u3Va27, i1Xb27, i2Xb27, i3Xb27, i1Va27, i2Va27, i3Va27, u1Xb28, u2Xb28, u3Xb28, u1Va28, u2Va28, u3Va28, i1Xb28, i2Xb28, i3Xb28, i1Va28, i2Va28, i3Va28, u1Xb29, u2Xb29, u3Xb29, u1Va29, u2Va29, u3Va29, i1Xb29, i2Xb29, i3Xb29, i1Va29, i2Va29, i3Va29, u1Xb30, u2Xb30, u3Xb30, u1Va30, u2Va30, u3Va30, i1Xb30, i2Xb30, i3Xb30, i1Va30, i2Va30, i3Va30, u1Xb31, u2Xb31, u3Xb31, u1Va31, u2Va31, u3Va31, i1Xb31, i2Xb31, i3Xb31, i1Va31, i2Va31, i3Va31, u1Xb32, u2Xb32, u3Xb32, u1Va32, u2Va32, u3Va32, i1Xb32, i2Xb32, i3Xb32, i1Va32, i2Va32, i3Va32, u1Xb33, u2Xb33, u3Xb33, u1Va33, u2Va33, u3Va33, i1Xb33, i2Xb33, i3Xb33, i1Va33, i2Va33, i3Va33, u1Xb34, u2Xb34, u3Xb34, u1Va34, u2Va34, u3Va34, i1Xb34, i2Xb34, i3Xb34, i1Va34, i2Va34, i3Va34, u1Xb35, u2Xb35, u3Xb35, u1Va35, u2Va35, u3Va35, i1Xb35, i2Xb35, i3Xb35, i1Va35, i2Va35, i3Va35, u1Xb36, u2Xb36, u3Xb36, u1Va36, u2Va36, u3Va36, i1Xb36, i2Xb36, i3Xb36, i1Va36, i2Va36, i3Va36, u1Xb37, u2Xb37, u3Xb37, u1Va37, u2Va37, u3Va37, i1Xb37, i2Xb37, i3Xb37, i1Va37, i2Va37, i3Va37, u1Xb38, u2Xb38, u3Xb38, u1Va38, u2Va38, u3Va38, i1Xb38, i2Xb38, i3Xb38, i1Va38, i2Va38, i3Va38, u1Xb39, u2Xb39, u3Xb39, u1Va39, u2Va39, u3Va39, i1Xb39, i2Xb39, i3Xb39, i1Va39, i2Va39, i3Va39, u1Xb40, u2Xb40, u3Xb40, u1Va40, u2Va40, u3Va40, i1Xb40, i2Xb40, i3Xb40, i1Va40, i2Va40, i3Va40, u1Xb41, u2Xb41, u3Xb41, u1Va41, u2Va41, u3Va41, i1Xb41, i2Xb41, i3Xb41, i1Va41, i2Va41, i3Va41, u1Xb42, u2Xb42, u3Xb42, u1Va42, u2Va42, u3Va42, i1Xb42, i2Xb42, i3Xb42, i1Va42, i2Va42, i3Va42, u1Xb43, u2Xb43, u3Xb43, u1Va43, u2Va43, u3Va43, i1Xb43, i2Xb43, i3Xb43, i1Va43, i2Va43, i3Va43, u1Xb44, u2Xb44, u3Xb44, u1Va44, u2Va44, u3Va44, i1Xb44, i2Xb44, i3Xb44, i1Va44, i2Va44, i3Va44, u1Xb45, u2Xb45, u3Xb45, u1Va45, u2Va45, u3Va45, i1Xb45, i2Xb45, i3Xb45, i1Va45, i2Va45, i3Va45, u1Xb46, u2Xb46, u3Xb46, u1Va46, u2Va46, u3Va46, i1Xb46, i2Xb46, i3Xb46, i1Va46, i2Va46, i3Va46, u1Xb47, u2Xb47, u3Xb47, u1Va47, u2Va47, u3Va47, i1Xb47, i2Xb47, i3Xb47, i1Va47, i2Va47, i3Va47, u1Xb48, u2Xb48, u3Xb48, u1Va48, u2Va48, u3Va48, i1Xb48, i2Xb48, i3Xb48, i1Va48, i2Va48, i3Va48, u1Xb49, u2Xb49, u3Xb49, u1Va49, u2Va49, u3Va49, i1Xb49, i2Xb49, i3Xb49, i1Va49, i2Va49, i3Va49, u1Xb50, u2Xb50, u3Xb50, u1Va50, u2Va50, u3Va50, i1Xb50, i2Xb50, i3Xb50, i1Va50, i2Va50, i3Va50);
+    }
+
+    public Float getThdu1() {
+        return thdu1;
+    }
+
+    public void setThdu1(Float thdu1) {
+        this.thdu1 = thdu1;
+    }
+
+    public Float getThdu2() {
+        return thdu2;
+    }
+
+    public void setThdu2(Float thdu2) {
+        this.thdu2 = thdu2;
+    }
+
+    public Float getThdu3() {
+        return thdu3;
+    }
+
+    public void setThdu3(Float thdu3) {
+        this.thdu3 = thdu3;
+    }
+
+    public Float getThdi1() {
+        return thdi1;
+    }
+
+    public void setThdi1(Float thdi1) {
+        this.thdi1 = thdi1;
+    }
+
+    public Float getThdi2() {
+        return thdi2;
+    }
+
+    public void setThdi2(Float thdi2) {
+        this.thdi2 = thdi2;
+    }
+
+    public Float getThdi3() {
+        return thdi3;
+    }
+
+    public void setThdi3(Float thdi3) {
+        this.thdi3 = thdi3;
     }
 }
