+++ src/onlineTest/dao/impl/RMSDAOImpl.java
@@ -34,149 +34,6 @@ public class RMSDAOImpl implements RMSDAO {
     double maxv1 = 0,maxv2 = 0,maxv3 = 0;
 
 
-   /* public void getCurrentData(String did) {
-        HBSessionDaoImpl hbsessionDao = new HBSessionDaoImpl();
-
-        u1list = hbsessionDao.searchWithNum(
-                "FROM PowerxbMonitor where did = '" + did + "' order by time desc", 5);
-
-        if (u1list.size() < 3) {
-            for (int i = 0; i < u1list.size(); i++) {
-                if (u1list.get(i).getU1Xb3() >= maxu1) {
-                    maxu1 = u1list.get(i).getU1Xb3();
-                }
-                if (u1list.get(i).getU2Xb3() >= maxu2) {
-                    maxu2 = u1list.get(i).getU2Xb3();
-                }
-                if (u1list.get(i).getU3Xb3() >= maxu3) {
-                    maxu3 = u1list.get(i).getU3Xb3();
-                }
-
-                if (u1list.get(i).getI1Xb3() >= maxi1) {
-                    maxi1 = u1list.get(i).getI1Xb3();
-                }
-                if (u1list.get(i).getI2Xb3() >= maxi2) {
-                    maxi2 = u1list.get(i).getI2Xb3();
-                }
-                if (u1list.get(i).getI3Xb3() >= maxi3) {
-                    maxi3 = u1list.get(i).getI3Xb3();
-                }
-
-                u1_1 += Math.pow(u1list.get(i).getU1Xb3(), 2);
-                u2_1 += Math.pow(u1list.get(i).getU2Xb3(), 2);
-                u3_1 += Math.pow(u1list.get(i).getU3Xb3(), 2);
-
-              *//*  v1_1 += Math.pow(u1list.get(i).getV1Xb3(), 2);
-                v2_1 += Math.pow(u1list.get(i).getV2Xb3(), 2);
-                v3_1 += Math.pow(u1list.get(i).getV3Xb3(), 2);*//*
-
-                i1_1 += Math.pow(u1list.get(i).getI1Xb3(), 2);
-                i2_1 += Math.pow(u1list.get(i).getI2Xb3(), 2);
-                i3_1 += Math.pow(u1list.get(i).getI3Xb3(), 2);
-            }
-        } else {
-            for (int j = 0; j < 3; j++) {
-               if (j == 0) {
-                   for (int i = 0; i < 3; i++) {
-                       if (u1list.get(i+j).getU1Xb3() >= maxu1) {
-                           maxu1 = u1list.get(i + j).getU1Xb3();
-                       }
-                       if (u1list.get(i+j).getU2Xb3() >= maxu2) {
-                           maxu2 = u1list.get(i + j).getU2Xb3();
-                       }
-                       if (u1list.get(i+j).getU3Xb3() >= maxu3) {
-                           maxu3 = u1list.get(i + j).getU3Xb3();
-                       }
-
-                       if(u1list.get(i+j).getI1Xb3() >= maxi1) {
-                           maxi1 = u1list.get(i + j).getI1Xb3();
-                       }
-                       if(u1list.get(i+j).getI2Xb3() >= maxi2) {
-                           maxi2 = u1list.get(i + j).getI2Xb3();
-                       }
-                       if(u1list.get(i+j).getI3Xb3() >= maxi3) {
-                           maxi3 = u1list.get(i + j).getI3Xb3();
-                       }
-
-                       u1_1 += Math.pow(u1list.get(i+j).getU1Xb3(), 2);
-                       u2_1 += Math.pow(u1list.get(i+j).getU2Xb3(), 2);
-                       u3_1 += Math.pow(u1list.get(i+j).getU3Xb3(), 2);
-             *//*   v1_1 += Math.pow(u1list.get(i).getV1Xb3(), 2);
-                v2_1 += Math.pow(u1list.get(i).getV2Xb3(), 2);
-                v3_1 += Math.pow(u1list.get(i).getV3Xb3(), 2);*//*
-                       i1_1 += Math.pow(u1list.get(i+j).getI1Xb3(), 2);
-                       i2_1 += Math.pow(u1list.get(i+j).getI2Xb3(), 2);
-                       i3_1 += Math.pow(u1list.get(i+j).getI3Xb3(), 2);
-                   }
-               } else if (j == 1) {
-                   for (int i = 0; i < 3; i++) {
-                       if (u1list.get(i+j).getU1Xb3() >= maxu1) {
-                           maxu1 = u1list.get(i + j).getU1Xb3();
-                       }
-                       if (u1list.get(i+j).getU2Xb3() >= maxu2) {
-                           maxu2 = u1list.get(i + j).getU2Xb3();
-                       }
-                       if (u1list.get(i+j).getU3Xb3() >= maxu3) {
-                           maxu3 = u1list.get(i + j).getU3Xb3();
-                       }
-
-                       if (u1list.get(i+j).getI1Xb3() >= maxi1) {
-                           maxi1 = u1list.get(i + j).getI1Xb3();
-                       }
-                       if (u1list.get(i+j).getI2Xb3() >= maxi2) {
-                           maxi2 = u1list.get(i + j).getI2Xb3();
-                       }
-                       if (u1list.get(i+j).getI3Xb3() >= maxi3) {
-                           maxi3 = u1list.get(i + j).getI3Xb3();
-                       }
-
-                       u1_2 += Math.pow(u1list.get(i+j).getU1Xb3(), 2);
-                       u2_2 += Math.pow(u1list.get(i+j).getU2Xb3(), 2);
-                       u3_2 += Math.pow(u1list.get(i+j).getU3Xb3(), 2);
-             *//*   v1_2 += Math.pow(u1list.get(i).getV1Xb3(), 2);
-                v2_2 += Math.pow(u1list.get(i).getV2Xb3(), 2);
-                v3_2 += Math.pow(u1list.get(i).getV3Xb3(), 2);*//*
-                       i1_2 += Math.pow(u1list.get(i+j).getI1Xb3(), 2);
-                       i2_2 += Math.pow(u1list.get(i+j).getI2Xb3(), 2);
-                       i3_2 += Math.pow(u1list.get(i+j).getI3Xb3(), 2);
-                   }
-               } else if(j == 2) {
-                   for (int i = 0; i < 3; i++) {
-                       if(u1list.get(i+j).getU1Xb3() >= maxu1) {
-                           maxu1 = u1list.get(i + j).getU1Xb3();
-                       }
-                       if(u1list.get(i+j).getU2Xb3() >= maxu2) {
-                           maxu2 = u1list.get(i + j).getU2Xb3();
-                       }
-                       if(u1list.get(i+j).getU3Xb3() >= maxu3) {
-                           maxu3 = u1list.get(i + j).getU3Xb3();
-                       }
-
-                       if(u1list.get(i+j).getI1Xb3() >= maxi1) {
-                           maxi1 = u1list.get(i + j).getI1Xb3();
-                       }
-                       if(u1list.get(i+j).getI2Xb3() >= maxi2) {
-                           maxi2 = u1list.get(i + j).getI2Xb3();
-                       }
-                       if(u1list.get(i+j).getI3Xb3() >= maxi3) {
-                           maxi3 = u1list.get(i + j).getI3Xb3();
-                       }
-
-                       u1_3 += Math.pow(u1list.get(i+j).getU1Xb3(), 2);
-                       u2_3 += Math.pow(u1list.get(i+j).getU2Xb3(), 2);
-                       u3_3 += Math.pow(u1list.get(i+j).getU3Xb3(), 2);
-             *//*   v1_3 += Math.pow(u1list.get(i).getV1Xb3(), 2);
-                v2_3 += Math.pow(u1list.get(i).getV2Xb3(), 2);
-                v3_3 += Math.pow(u1list.get(i).getV3Xb3(), 2);*//*
-                       i1_3 += Math.pow(u1list.get(i+j).getI1Xb3(), 2);
-                       i2_3 += Math.pow(u1list.get(i+j).getI2Xb3(), 2);
-                       i3_3 += Math.pow(u1list.get(i+j).getI3Xb3(), 2);
-                   }
-               }
-            }
-        }
-    }*/
-
     public List getCurrentRMSData(String did) {
         List<RMS> crlist = new ArrayList<>();
         HBSessionDaoImpl hbsessionDao = new HBSessionDaoImpl();
@@ -205,8 +62,8 @@ public class RMSDAOImpl implements RMSDAO {
         List<THD> crlist = new ArrayList<>();
         HBSessionDaoImpl hbsessionDao = new HBSessionDaoImpl();
 
-        List<PowerparmMonitor> obj = hbsessionDao.searchWithNum(
-                "FROM PowerparmMonitor where did = '" + did + "' order by time desc", 1);
+        List<PowerxbMonitor> obj = hbsessionDao.searchWithNum(
+                "FROM PowerxbMonitor where did = '" + did + "' order by time desc", 1);
 
         for (int i = 0; i < obj.size(); i++) {
             THD rtobj = new THD();
